third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/calc-size-width-interpolation.tentative-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [150px]
  assert_equals: expected "150px " but got "400px "
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (0) should be [200px]
  assert_equals: expected "200px " but got "400px "
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [250px]
  assert_equals: expected "250px " but got "400px "
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [300px]
  assert_equals: expected "300px " but got "400px "
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [350px]
  assert_equals: expected "350px " but got "400px "
[FAIL] CSS Transitions: property <width> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [450px]
  assert_equals: expected "450px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [150px]
  assert_equals: expected "150px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (0) should be [200px]
  assert_equals: expected "200px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [250px]
  assert_equals: expected "250px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [300px]
  assert_equals: expected "300px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [350px]
  assert_equals: expected "350px " but got "400px "
[FAIL] CSS Transitions with transition: all: property <width> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [450px]
  assert_equals: expected "450px " but got "400px "
[FAIL] CSS Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] CSS Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [250px]
  assert_equals: expected "250px " but got "200px "
[FAIL] CSS Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [300px]
  assert_equals: expected "300px " but got "400px "
[FAIL] CSS Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [350px]
  assert_equals: expected "350px " but got "400px "
[FAIL] CSS Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [450px]
  assert_equals: expected "450px " but got "400px "
[FAIL] Web Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [150px]
  assert_equals: expected "150px " but got "200px "
[FAIL] Web Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [250px]
  assert_equals: expected "250px " but got "200px "
[FAIL] Web Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [300px]
  assert_equals: expected "300px " but got "400px "
[FAIL] Web Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [350px]
  assert_equals: expected "350px " but got "400px "
[FAIL] Web Animations: property <width> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [450px]
  assert_equals: expected "450px " but got "400px "
Harness: the test ran to completion.
